refactor(core,negotiation,prompt): register stray error codes, consolidate duplicate value types and align nullness contracts
