--- a2a-t-core/src/main/java/net/openan/a2at/sdk/core/exception/FailedParameter.java
@@ -1,12 +1,0 @@
-package net.openan.a2at.sdk.core.exception;
-
-import org.jspecify.annotations.NonNull;
-
-/**
- * Describes a single failed required slot during slot validation.
- *
- * @param parameterName the name of the required slot that failed validation
- * @param reason the failure reason, e.g. {@code missing_required}
- * @since 2026-08
- */
-public record FailedParameter(@NonNull String parameterName, @NonNull String reason) {}
