+++ a2a-t-negotiation/src/main/java/net/openan/a2at/sdk/negotiation/validation/NegotiationComplianceChecker.java
@@ -1,7 +1,5 @@
 package net.openan.a2at.sdk.negotiation.validation;
 
-import java.util.Map;
-import net.openan.a2at.sdk.core.validation.RuleChecker;
 import net.openan.a2at.sdk.negotiation.content.Vocabulary;
 
 /**
@@ -16,7 +14,7 @@ import net.openan.a2at.sdk.negotiation.content.Vocabulary;
  *
  * @since 2026-06
  */
-public interface NegotiationComplianceChecker extends RuleChecker {
+public interface NegotiationComplianceChecker {
 
     /**
      * Runs the rule-level compliance check of one rendered message.
@@ -30,14 +28,4 @@ public interface NegotiationComplianceChecker extends RuleChecker {
      */
     NegotiationRuleCheckResult check(String prompt, Vocabulary vocabulary);
 
-    /**
-     * {@inheritDoc}
-     *
-     * <p>This default implementation is not supported; use {@link #check(String, Vocabulary)} instead. The
-     * {@link NegotiationRuleCheckerAdapter} bridges this interface to the {@link RuleChecker} contract.
-     */
-    @Override
-    default Map<String, Object> check(String prompt) {
-        throw new UnsupportedOperationException("Use check(String, Vocabulary) instead");
-    }
 }
