+++ a2a-t-server/src/main/java/net/openan/a2at/sdk/server/validation/LlmBackedPromptSemanticValidator.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import net.openan.a2at.sdk.core.model.PromptMessage;
+import net.openan.a2at.sdk.core.exception.A2ATErrorCodes;
 import net.openan.a2at.sdk.llm.LLMClient;
 import net.openan.a2at.sdk.prompt.resources.loader.PromptSlotSchemaLoader;
 import net.openan.a2at.sdk.prompt.resources.model.PromptSlotSchema;
@@ -74,7 +75,7 @@ public final class LlmBackedPromptSemanticValidator implements ServerPromptSeman
                     + "\n}";
         } catch (JsonProcessingException error) {
             throw new PromptComplianceCheckException(
-                    "slot_validation_error", "Failed to serialize slot schema", "slot_validation");
+                    A2ATErrorCodes.SLOT_VALIDATION_ERROR, "Failed to serialize slot schema", "slot_validation");
         }
     }
 
@@ -118,7 +119,7 @@ public final class LlmBackedPromptSemanticValidator implements ServerPromptSeman
 
         Optional<String> message = extractFirstMessage(errorsValue);
         throw new PromptComplianceCheckException(
-                "slot_validation_error",
+                A2ATErrorCodes.SLOT_VALIDATION_ERROR,
                 message.filter(text -> !text.isBlank()).orElse("Slot semantic validation failed."),
                 "slot_validation");
     }
@@ -130,7 +131,7 @@ public final class LlmBackedPromptSemanticValidator implements ServerPromptSeman
             return Optional.ofNullable(response).orElseGet(Map::of);
         } catch (JsonProcessingException error) {
             throw new PromptComplianceCheckException(
-                    "slot_validation_error", "semantic validation returned invalid JSON", "slot_validation");
+                    A2ATErrorCodes.SLOT_VALIDATION_ERROR, "semantic validation returned invalid JSON", "slot_validation");
         }
     }
